Update README.md
The originally linked ipfs-userscript had problems: the first version was redirecting to https-localhost instead of http-localhost, i.e. not working, and the current version is kinda working, but it doesn't support complex ipfs/ipns paths; so I'm offering a new redirect script that should (hopefully!) work; it also checks for a local ipfs node and only redirects, if it's actually running. The script should also move to "Tools", where the browser addons/extensions are.