Update pull_request_template.md
First pass at an improved PR template for awesome-ipfs. @andrew -- we'll need to change the pre-submit checklist based on what you figure out surrounding the mechanisms of making a PR easier to make. In the meantime, please take a look at the remaining questions to see if they're useful for getting a bead on collabs.

cc @momack2 @autonome @renrutnnej